+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.18] - 2022-09-26
+### Added
+- Compendium items.
+- Give Ground status effect.
+- Scamper Back status effect.
+### Changed
+- Magic always uses combat encounter's round for timing purposes.
+- Spell chat cards are more explicit about weapon resets.
+
 ## [0.2.17] - 2022-09-09
 ### Added
 - Compendium items.